chore: Move theme related helpers to a new file
